Move service to new class
